# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ayo Indonesia Api",
        "description": "Ayo Indonesia Api Documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/teams": {
            "get": {
                "tags": [
                    "Teams"
                ],
                "summary": "get list of teams",
                "description": "Returns list of teams",
                "operationId": "getTeamList",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Teams"
                ],
                "summary": "create new team",
                "description": "create data team",
                "operationId": "createteam",
                "requestBody": {
                    "description": "create team",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "logo",
                                    "since",
                                    "address",
                                    "city_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Persebaya"
                                    },
                                    "logo": {
                                        "type": "file",
                                        "format": "file",
                                        "example": ""
                                    },
                                    "since": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1973-08-05"
                                    },
                                    "address": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Glora Bung Tomo Surabaya"
                                    },
                                    "city_id": {
                                        "type": "number",
                                        "format": "number",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successfully Created"
                    },
                    "422": {
                        "description": "Unprocessable Content - Validation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/teams/{id}": {
            "get": {
                "tags": [
                    "Teams"
                ],
                "summary": "get team by id",
                "description": "get team by id",
                "operationId": "getTeamById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "get team by id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Teams"
                ],
                "summary": "update team by id",
                "description": "update team data",
                "operationId": "updateteamById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "update team by id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "update team by id",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "logo": {
                                        "type": "file",
                                        "format": "file",
                                        "example": ""
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Persebaya"
                                    },
                                    "since": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1973-08-05"
                                    },
                                    "address": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Glora Bung Tomo Surabaya"
                                    },
                                    "city_id": {
                                        "type": "number",
                                        "format": "number",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "422": {
                        "description": "Unprocessable Content - Validation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Teams"
                ],
                "summary": "delete team by id",
                "description": "delete team by Id",
                "operationId": "deleteTeamById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "delete team by id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Success - No Content"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}